% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aritmetica.R
\name{aritmetica}
\alias{aritmetica}
\title{Operacoes Matematicas Basicas}
\usage{
aritmetica(a, b, op)
}
\arguments{
\item{a}{primeiro numero}

\item{b}{segundo numero}

\item{op}{operacao matematica}
}
\value{
A funcao \code{aritmetica} realiza as quatro operacoes basicas.
    A opcoes para o argumento \code{op} sao "soma", "divisi",
    "multi", "subtra"
}
\description{
\code{aritmetica} resolve operacoes como: soma, subtracao,
    divisao e multiplicacao.
}
\examples{
# Loading package
library(meupacoter)
aritmetica(8, 4, op = "divisi")
aritmetica(8, 4, op = "soma")
aritmetica(8, 4, op = "subtra")
aritmetica(8, 4, op = "multi")
}
